Move code to seperate file and reduce dot count in left bar to 1 of each type
