+++ app/javascript/packs/submission.js
@@ -1,5 +1,5 @@
 import { initSubmissionShow } from "submission.js";
-import { CodeListing } from "code_listing.ts";
+import { CodeListing } from "code_listing/code_listing.ts";
 
 window.dodona.initSubmissionShow = initSubmissionShow;
 window.dodona.codeListingClass = CodeListing;
